back/requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
yoyo-migrations==8.2.0
SQLAlchemy==2.0.17
psycopg2==2.9.6
PyJWT==2.7.0
python-multipart~=0.0.6
pydantic==1.10.9
opencv-python==4.7.0.72
coverage==7.3.2
httpx==0.25.2
pytest-cov==4.1.0
locust==2.19.1
openai~=0.28.1
